api/requirements.txt
# requirements.txt
fastapi
uvicorn
langchain-core
langchain_community
langchain_chroma
docx2txt
pypdf
python-multipart
#sqlite3
ollama
python-dotenv
# sentence-transformers # Uncomment if you use HuggingFace embeddings instead of Ollamacls
